requirements.txt
streamlit==1.28.2
openai
google-generativeai
embedchain
chromadb
python-dotenv
diffusers==0.10.2
transformers 
scipy 
ftfy 
accelerate
torchaudio
torch
torchvision
